+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project that require documentation updates will be d
 ## [NEXT RELEASE]
 - Detection APIs were not properly handling suppressed CVEs and they were being included in evaluation. This is now resolved.
 - Previously, the Scanner deployment did not mount the additional CA secret and thus would fail to scan self-signed registries. This is resolved.
+- AWS S3 and AWS ECR integrations now accept an endpoint to work with non public AWS endpoints.
 
 ## [42.0]
 - All `/v1/` API endpoints now support pretty-printing.  Make requests with the `?pretty` path parameter to receive pretty-printed json responses.